move makeA1Column function
